refactor(cast): inline selector implementation
