+++ crates/cast/tests/cli/main.rs
@@ -9917,3 +9917,7 @@ casttest!(disassemble_incomplete_sequence, |_prj, cmd| {
         .stdout_eq("00000000: PUSH32\n\n");
     cmd.cast_fuse().args(["disassemble", "7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff"]).assert_success().stdout_eq("00000000: PUSH32 0xffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff\n\n");
 });
+
+casttest!(signature_selector, |_prj, cmd| {
+    cmd.cast_fuse().args(["sig", "foo()"]).assert_success().stdout_eq("0xc2985578\n");
+});
